added new test script, logging

+++ scripts/init.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-printf "***** Compiling kong middleware *****\n\n"
+echo "*************************************"
+echo "Compiling kong middleware"
+echo "*************************************"
 
 cd /kong-plugin
 
@@ -16,14 +18,20 @@ for i in *; do
   fi
 done
 
-printf "***** Compiling kong *****\n\n"
+echo "*************************************"
+echo "Compiling kong"
+echo "*************************************"
 
 cd /
 make --directory=./kong dev
 
-printf "***** installing terraform *****\n\n"
+if [ ! -f /usr/local/bin/terraform ] ; then
+  echo "*************************************"
+  echo "Installing terraform"
+  echo "*************************************"
 
-cd /tmp
-wget --quiet https://releases.hashicorp.com/terraform/0.12.17/terraform_0.12.17_linux_amd64.zip
-unzip -qq terraform*.zip
-mv ./terraform /usr/local/bin/
+  cd /tmp
+  wget --quiet https://releases.hashicorp.com/terraform/0.12.17/terraform_0.12.17_linux_amd64.zip
+  unzip -qq terraform*.zip
+  mv ./terraform /usr/local/bin/
+fi
